BVAL-526 Moving project to use maven
