+++ .gitignore
@@ -6,6 +6,10 @@
 .classpath
 .settings
 
+#Idea
+.idea
+*.iml
+
 #Build
 lib/
 target/
