+++ documentation/src/main/asciidoc/appendix-standardresolvermessages.asciidoc
