+++ documentation/src/main/asciidoc/appendix-value-extraction.asciidoc
