+++ documentation/src/main/asciidoc/constraint-metadata.asciidoc
