+++ documentation/src/main/asciidoc/exception.asciidoc
