+++ documentation/src/main/asciidoc/index.asciidoc
@@ -1,0 +1,66 @@
+// Bean Validation
+//
+// License: Apache License, Version 2.0
+// See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
+= Bean Validation specification
+Gunnar Morling
+:revnumber: {bv-version}{bv-version-qualifier}
+:revdate: {bv-revdate}
+:sectanchors:
+:anchor:
+:toc: left
+:toclevels: 3
+:sectnumlevels: 5
+:docinfo:
+:docinfodir: ../docinfo
+:title-logo-image: image:beanvalidation_logo.png[align=left,pdfwidth=20%]
+
+[preface]
+include::{license}.asciidoc[]
+
+:sectnums:
+
+<<<
+include::introduction.asciidoc[]
+
+<<<
+include::whatsnew.asciidoc[]
+
+<<<
+include::constraint-definition.asciidoc[]
+
+<<<
+include::constraint-declaration-validation.asciidoc[]
+
+<<<
+include::validation-api.asciidoc[]
+
+<<<
+include::constraint-metadata.asciidoc[]
+
+<<<
+include::builtin-constraints.asciidoc[]
+
+<<<
+include::xml-descriptor.asciidoc[]
+
+<<<
+include::exception.asciidoc[]
+
+<<<
+include::integration.asciidoc[]
+
+<<<
+include::terminology.asciidoc[]
+
+<<<
+include::appendix-value-extraction.asciidoc[]
+
+<<<
+include::appendix-standardresolvermessages.asciidoc[]
+
+<<<
+include::appendix-jpa.asciidoc[]
+
+<<<
+include::changelog.asciidoc[]
