+++ documentation/src/main/asciidoc/xml-descriptor.asciidoc
