--- sources/index.asciidoc
@@ -1,66 +1,0 @@
-// Bean Validation
-//
-// License: Apache License, Version 2.0
-// See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-= Bean Validation specification
-Gunnar Morling
-:revnumber: {bv-version}{bv-version-qualifier}
-:revdate: {bv-revdate}
-:sectanchors:
-:anchor:
-:toc: left
-:toclevels: 3
-:sectnumlevels: 5
-:docinfo:
-:docinfodir: ../docinfo
-:title-logo-image: image:beanvalidation_logo.png[align=left,pdfwidth=20%]
-
-[preface]
-include::sources/{license}.asciidoc[]
-
-:sectnums:
-
-<<<
-include::sources/introduction.asciidoc[]
-
-<<<
-include::sources/whatsnew.asciidoc[]
-
-<<<
-include::sources/constraint-definition.asciidoc[]
-
-<<<
-include::sources/constraint-declaration-validation.asciidoc[]
-
-<<<
-include::sources/validation-api.asciidoc[]
-
-<<<
-include::sources/constraint-metadata.asciidoc[]
-
-<<<
-include::sources/builtin-constraints.asciidoc[]
-
-<<<
-include::sources/xml-descriptor.asciidoc[]
-
-<<<
-include::sources/exception.asciidoc[]
-
-<<<
-include::sources/integration.asciidoc[]
-
-<<<
-include::sources/terminology.asciidoc[]
-
-<<<
-include::sources/appendix-value-extraction.asciidoc[]
-
-<<<
-include::sources/appendix-standardresolvermessages.asciidoc[]
-
-<<<
-include::sources/appendix-jpa.asciidoc[]
-
-<<<
-include::sources/changelog.asciidoc[]
